Add RatioGrader block (#80)
We add the RatioGrader block, with tests. This is mostly a test fixture for continuing to develop CapaProblem, which has issues right now. A lot is done in the parser which should be done in the renderer, and we don't support multiple graders in the same problem. To do development with meaningful testing, we do need an example grader which has multiple inputs.