Comprehensive documentation cleanup (#146)
Comprehensive rewrite of documentation. This provides a simple overview file (usually, usable as a template in the editor), and further examples and documentation in back. Playgrounds in the readme files. Symlinks where relevant. 

We tried to tie examples to education (but we could still do a more thoughtful job).

Along the way, we ran into plenty of bugs in the system, which we mostly fixed, or in some cases, documented. Plenty of work left, but a significant step up in quality.